+++ internal/api/migrate/migrate.go
@@ -1,4 +1,4 @@
-// Copyright 2023 The Perses Authors
+// Copyright 2024 The Perses Authors
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
 // You may obtain a copy of the License at
@@ -14,56 +14,21 @@
 package migrate
 
 import (
-	_ "embed"
 	"encoding/json"
-	"errors"
 	"fmt"
 	"strconv"
 	"strings"
-	"sync/atomic"
+	"time"
 
 	"cuelang.org/go/cue"
+	"cuelang.org/go/cue/build"
 	"cuelang.org/go/cue/cuecontext"
 	apiinterface "github.com/perses/perses/internal/api/interface"
-	"github.com/perses/perses/internal/api/schemas"
 	"github.com/perses/perses/pkg/model/api/config"
 	v1 "github.com/perses/perses/pkg/model/api/v1"
+	"github.com/perses/perses/pkg/model/api/v1/common"
+	"github.com/perses/perses/pkg/model/api/v1/dashboard"
 	"github.com/sirupsen/logrus"
-)
-
-//go:embed mapping.cuepart
-var migrationFileBytes []byte
-
-const (
-	grafanaDefID = "#grafanaDashboard"
-
-	variableDefaultValue = `
-		kind: "StaticListVariable"
-		spec: {
-			values: ["grafana", "migration", "not", "supported"]
-		}
-	`
-	variablePlaceholderText = "%(conditional_variables)"
-
-	panelDefaultValue = `
-		kind: "Markdown"
-		spec: {
-			text: "**Migration from Grafana not supported !**"
-		}
-	`
-	panelPlaceholderText = "%(conditional_panels)"
-
-	queryDefaultValue = `
-		kind: "PrometheusTimeSeriesQuery"
-		spec: {
-			datasource: {
-				kind: "PrometheusDatasource"
-				name: "MigrationFromGrafanaNotSupported"
-			}
-			query: "migration_from_grafana_not_supported"
-		}
-	`
-	queryPlaceholderText = "%(conditional_timeseries_queries)"
 )
 
 func ReplaceInputValue(input map[string]string, grafanaDashboard string) string {
@@ -81,250 +46,160 @@ func ReplaceInputValue(input map[string]string, grafanaDashboard string) string 
 }
 
 type Migration interface {
-	Migrate(grafanaDashboard []byte) (*v1.Dashboard, error)
-	BuildMigrationSchemaString()
-	GetLoaders() []schemas.Loader
+	Migrate(grafanaDashboard *SimplifiedDashboard) (*v1.Dashboard, error)
 }
 
-func New(schemasConf config.Schemas) (Migration, error) {
-	m := &mig{
-		loaders: []loader{
-			&migCuePart{
-				schemasPath:     schemasConf.VariablesPath,
-				defaultValue:    variableDefaultValue,
-				placeholderText: variablePlaceholderText,
-			},
-			&migCuePart{
-				schemasPath:     schemasConf.PanelsPath,
-				defaultValue:    panelDefaultValue,
-				placeholderText: panelPlaceholderText,
-			},
-			&migCuePart{
-				schemasPath:     schemasConf.QueriesPath,
-				defaultValue:    queryDefaultValue,
-				placeholderText: queryPlaceholderText,
-			},
-		},
-	}
-	if err := m.init(); err != nil {
+func New(schemas config.Schemas) (Migration, error) {
+	panels, err := loadPanels(schemas.PanelsPath)
+	if err != nil {
 		return nil, err
 	}
-	return m, nil
+	queries, err := loadSliceOfInstance(schemas.QueriesPath)
+	if err != nil {
+		return nil, err
+	}
+	variables, err := loadSliceOfInstance(schemas.VariablesPath)
+	return &mig{
+		panels:    panels,
+		queries:   queries,
+		variables: variables,
+	}, nil
 }
 
 type mig struct {
-	migrationSchemaString atomic.Pointer[string]
-	loaders               []loader
+	// panels is a map because we can decide which script to execute precisely.
+	// This is because in Grafana a panel has a type.
+	panels map[string]*build.Instance
+	// Dynamic variables are usually in a parameter named 'query' in the Grafana data model and depending on what contains the query, then it will change the type of the plugin.
+	// That would mean we would have to parse the string to know what script to use. Which means hardcoding things which is against the plugin philosophy.
+	variables []*build.Instance
+	queries   []*build.Instance
 }
 
-func (m *mig) GetLoaders() []schemas.Loader {
-	var loaders []schemas.Loader
-	for _, l := range m.loaders {
-		loaders = append(loaders, l)
-	}
-	return loaders
-}
-
-func (m *mig) BuildMigrationSchemaString() {
-	// start building the migration schema from the base mapping.cuepart file
-	migrationSchemaString := string(migrationFileBytes)
-
-	// generate the blocks of conditionals from the migrate.cue files of each plugin & replace the placeholders with them
-	for _, l := range m.loaders {
-		conditionals := l.getConditions()
-		migrationSchemaString = strings.Replace(migrationSchemaString, l.getPlaceholder(), conditionals, -1)
-	}
-	logrus.Tracef("migrationSchemaString: %s", migrationSchemaString)
-	m.migrationSchemaString.Store(&migrationSchemaString)
-}
-
-func (m *mig) Migrate(userInput []byte) (*v1.Dashboard, error) {
-	cueContext := cuecontext.New(cuecontext.EvaluatorVersion(cuecontext.EvalV3))
-	// preprocessing in Go before passing it to CUE
-	grafanaDashboard := rearrangeGrafanaPanelsWithinExpandedRows(userInput)
-
-	// Build a CUE value that is just the received grafana dashboard wrapped in a definition. By doing this we isolate the
-	// fields from the Grafana dashboard together in a common namespace, so that they are not mixed with the ones from the
-	// migration schema. This make sure we have no wrong overlapping between Grafana & Perses objects, and also makes the
-	// remapping operations more explicit, with e.g `name: #grafanaDashboard.title` instead of `name: title`.
-	grafanaDashboardVal := cueContext.CompileString(fmt.Sprintf("%s: _", grafanaDefID))
-	grafanaDashboardVal = grafanaDashboardVal.FillPath(
-		cue.ParsePath(grafanaDefID),
-		cueContext.CompileBytes(grafanaDashboard),
-	)
-	if err := grafanaDashboardVal.Validate(cue.Final()); err != nil {
-		logrus.WithError(err).Trace("Unable to wrap the received json into a CUE definition")
-		return nil, apiinterface.HandleBadRequestError(err.Error())
+func (m *mig) Migrate(grafanaDashboard *SimplifiedDashboard) (*v1.Dashboard, error) {
+	result := &v1.Dashboard{
+		Kind: v1.KindDashboard,
+		Metadata: v1.ProjectMetadata{
+			Metadata: v1.Metadata{
+				Name: grafanaDashboard.UID,
+			},
+		},
+		Spec: v1.DashboardSpec{
+			Display: &common.Display{
+				Name: grafanaDashboard.Title,
+			},
+			Duration: common.Duration(time.Hour),
+		},
 	}
 
-	// Compile the migration schema using the grafana def to resolve the paths
-	mappingVal := cueContext.CompileString(*m.migrationSchemaString.Load(), cue.Scope(grafanaDashboardVal))
-	err := mappingVal.Err()
+	panels, err := m.migratePanels(grafanaDashboard)
 	if err != nil {
-		logrus.WithError(err).Trace("Unable to compile the migration schema using the received dashboard to resolve the paths")
-		return nil, apiinterface.HandleBadRequestError(fmt.Sprintf("unable to convert to Perses dashboard: %s", err))
+		return nil, err
 	}
-	logrus.Tracef("final migration schema: %#v", mappingVal)
-
-	// marshall to JSON then unmarshall in v1.Dashboard struct to pass the final checks & build the final dashboard to return
-	persesDashboardJSON, err := json.Marshal(mappingVal)
+	result.Spec.Panels = panels
+	variables, err := m.migrateVariables(grafanaDashboard)
 	if err != nil {
-		return nil, fmt.Errorf("%w: Unable to marshall CUE Value to json: %s", apiinterface.InternalError, err)
+		return nil, err
 	}
-	logrus.Tracef("perses dashboard as JSON: %s", string(persesDashboardJSON))
-
-	var persesDashboard v1.Dashboard
-	err = json.Unmarshal(persesDashboardJSON, &persesDashboard)
-	if err != nil {
-		logrus.WithError(err).Trace("Unable to unmarshall JSON bytes to Dashboard struct")
-		return nil, apiinterface.HandleBadRequestError(fmt.Sprintf("the Perses dashboard constraints are not met: %s", err))
-	}
-
-	return &persesDashboard, nil
+	result.Spec.Variables = variables
+	result.Spec.Layouts = m.migrateGrid(grafanaDashboard)
+	return result, nil
 }
 
-func (m *mig) init() error {
-	err := schemas.RunLoaders(m.GetLoaders(), schemas.Migration, schemas.Full)
-	if err != nil {
-		return err
-	}
-	m.BuildMigrationSchemaString()
-	return nil
-}
-
-/**
- * This function addresses an issue we have with Grafana datamodel when it comes to migrating dashboards to Perses: When
- * a row is expanded in Grafana, its children panels are moved up in the main panels list, thus become siblings of the row.
- * When it comes to Perses migration we need to recompose the parent->children relationships. However in its current state
- * the CUE language doesn't permit us to achieve this recomposition, hence this processing in the backend code.
- *
- * So what this function does is basically the following: whenever such pattern is encountered in the panels list:
- * ...
- * row1,
- * panelA,
- * panelB,
- * panelC,
- * row2,
- * ...
- * the objects gets rearranged like:
- * ...
- * row1: {
- *   panelA,
- *   panelB,
- *   panelC,
- * },
- * row2,
- * ...
- */
-func rearrangeGrafanaPanelsWithinExpandedRows(grafanaDashboardRaw json.RawMessage) json.RawMessage {
-	// unmarshall the received raw Json
-	var grafanaDashboard map[string]any
-	if err := json.Unmarshal(grafanaDashboardRaw, &grafanaDashboard); err != nil {
-		logrus.WithError(err).Error("Unable to unmarshall the received Grafana dashboard")
-		return grafanaDashboardRaw
+func (m *mig) migrateGrid(grafanaDashboard *SimplifiedDashboard) []dashboard.Layout {
+	var result []dashboard.Layout
+	defaultSpec := &dashboard.GridLayoutSpec{}
+	defaultLayout := dashboard.Layout{
+		Kind: dashboard.KindGridLayout,
+		Spec: defaultSpec,
 	}
 
-	// retrieve the panels
-	if _, found := grafanaDashboard["panels"]; !found {
-		logrus.Error(errors.New("expected `panels` field not found"))
-		return grafanaDashboardRaw
-	}
-	panels, ok := grafanaDashboard["panels"].([]any)
-	if !ok {
-		logrus.Error(errors.New("failed to assert `panels` to array of any"))
-		return grafanaDashboardRaw
-	}
+	result = append(result, defaultLayout)
 
-	// iterate over the panels & achieve recomposition of parent->children relationship when needed
-	newPanelList := make([]map[string]any, 0)
-	var parentRow map[string]any
-	for _, panelAsAny := range panels {
-		panel, ok := panelAsAny.(map[string]any)
-		if !ok {
-			logrus.Error(errors.New("failed to assert current panel to map of any"))
-			return grafanaDashboardRaw
-		}
-
-		if panel["type"] == "row" {
-			if _, found := panel["collapsed"]; !found {
-				// add collaped=false in row if missing
-				panel["collapsed"] = false
-			}
-			collapsed, ok := panel["collapsed"].(bool)
-			if !ok {
-				logrus.Error(errors.New("failed to assert the row's `collapsed` field to bool"))
-				return grafanaDashboardRaw
-			}
-
-			if parentRow != nil {
-				// situation corresponding to this case:
-				// row1,   <- current parentRow
-				// panelA,
-				// panelB,
-				// panelC,
-				// row2,   <- current iterated panel
-				// ...
-				// -> in this case, we should stop appending panels to the previously-registered parentRow,
-				// because we encountered a new row (we don't care if it is expanded or collapsed), thus we
-				// append parentRow to our new panel list. We also reset its value afterwards (parentRow will
-				// eventually be set to the newly-encountered row if it matches the expanded condition below)
-				newPanelList = append(newPanelList, parentRow)
-				parentRow = nil
-			}
-
-			if collapsed {
-				// any collapsed row should be appended as-is to our new panel list, without modifications.
-				newPanelList = append(newPanelList, panel)
-			} else {
-				// in this case we save the newly-encountered expanded row for the next iteration(s), since
-				// it's expanded.  We'll eventually have to append the next panel within it.
-				parentRow = panel
-			}
+	for i, panel := range grafanaDashboard.Panels {
+		if panel.Type != grafanaPanelRowType {
+			defaultSpec.Items = append(defaultSpec.Items, dashboard.GridItem{
+				Width:  panel.GridPosition.Width,
+				Height: panel.GridPosition.Height,
+				X:      panel.GridPosition.X,
+				Y:      panel.GridPosition.Y,
+				Content: &common.JSONRef{
+					Ref:  fmt.Sprintf("#/spec/panels/%d", i),
+					Path: []string{"spec", "panels", fmt.Sprintf("%d", i)},
+				},
+			})
 		} else {
-			if parentRow != nil {
-				// situation corresponding to this case:
-				// row1,   <- current parentRow
-				// panelA,
-				// panelB, <- current iterated panel
-				// ...
-				// -> in this case we have to move this non-row panel inside the saved parentRow
-				// add empty panels array to row if missing
-				if _, found := parentRow["panels"]; !found {
-					parentRow["panels"] = []any{}
-				}
-				subPanelsList, ok := parentRow["panels"].([]any)
-				if !ok {
-					logrus.Error(errors.New("failed to assert the row's `panels` field to array of any"))
-					return grafanaDashboardRaw
-				}
-				parentRow["panels"] = append(subPanelsList, panel)
-			} else {
-				// situation corresponding to this case:
-				// panelA,
-				// panelB, <- current iterated panel
-				// row1,
-				// ...
-				// -> in this case we append the panel as-is. Technically this case applies only when there are panels
-				// placed before any row
-				newPanelList = append(newPanelList, panel)
+			spec := &dashboard.GridLayoutSpec{
+				Display: &dashboard.GridLayoutDisplay{
+					Title: panel.Title,
+					Collapse: &dashboard.GridLayoutCollapse{
+						Open: !panel.Collapsed,
+					},
+				},
+			}
+			layout := dashboard.Layout{
+				Kind: dashboard.KindGridLayout,
+				Spec: spec,
+			}
+			for j, innerPanel := range panel.Panels {
+				spec.Items = append(spec.Items, dashboard.GridItem{
+					Width:  innerPanel.GridPosition.Width,
+					Height: innerPanel.GridPosition.Height,
+					X:      innerPanel.GridPosition.X,
+					Y:      innerPanel.GridPosition.Y,
+					Content: &common.JSONRef{
+						Ref:  fmt.Sprintf("#/spec/panels/%d_%d", i, j),
+						Path: []string{"spec", "panels", fmt.Sprintf("%d_%d", i, j)},
+					},
+				})
+			}
+			if len(spec.Items) > 0 {
+				result = append(result, layout)
 			}
 		}
 	}
-	// once the loop is over, it's possible that the last row we iterated over was expanded, but the loop finished,
-	// thus we need to append it here
-	if parentRow != nil {
-		newPanelList = append(newPanelList, parentRow)
+	if len(defaultSpec.Items) == 0 {
+		// Since there are no items, we should remove the default layout
+		result = result[1:]
+	}
+	return result
+}
+
+func executeCuelangMigrationScript(cueScript *build.Instance, grafanaData []byte, defID string, typeOfDataToMigrate string) (*common.Plugin, bool, error) {
+	ctx := cuecontext.New(cuecontext.EvaluatorVersion(cuecontext.EvalV3))
+	grafanaValue := ctx.CompileString(fmt.Sprintf("%s: _", defID))
+	grafanaValue = grafanaValue.FillPath(
+		cue.ParsePath(defID),
+		ctx.CompileBytes(grafanaData),
+	)
+
+	// Probably it is unnecessary to do that as JSON should be valid.
+	// Otherwise, we won't be able to unmarshal the grafana dashboard.
+	if err := grafanaValue.Validate(cue.Final()); err != nil {
+		logrus.WithError(err).Trace("Unable to wrap the received json into a CUE definition")
+		return nil, true, apiinterface.HandleBadRequestError(err.Error())
 	}
 
-	// overwrite the previous list of panels with the modified one & return:
+	// Finally, execute the cuelang script with the static mapping and the Grafana Panel as a scope.
+	finalVal := grafanaValue.Unify(ctx.BuildInstance(cueScript))
+	if err := finalVal.Err(); err != nil {
+		logrus.WithError(err).Debugf("Unable to compile the migration schema for the %s", typeOfDataToMigrate)
+		return nil, true, apiinterface.HandleBadRequestError(fmt.Sprintf("unable to convert to Perses %s: %s", typeOfDataToMigrate, err))
+	}
+	return convertToPlugin(finalVal)
+}
 
-	grafanaDashboard["panels"] = newPanelList
-
-	newGrafanaDashboardRaw, err := json.Marshal(grafanaDashboard)
+func convertToPlugin(migrateValue cue.Value) (*common.Plugin, bool, error) {
+	if migrateValue.IsNull() {
+		return nil, true, nil
+	}
+	data, err := migrateValue.MarshalJSON()
 	if err != nil {
-		logrus.WithError(err).Error("Unable to marshal the modified dashboard")
-		return grafanaDashboardRaw
+		return nil, true, err
 	}
-
-	return newGrafanaDashboardRaw
+	if string(data) == "" || string(data) == "{}" {
+		return nil, true, nil
+	}
+	plugin := &common.Plugin{}
+	return plugin, false, json.Unmarshal(data, plugin)
 }
