+++ internal/api/migrate/migrate_test.go
@@ -1,33 +1,27 @@
-// Copyright 2023 The Perses Authors
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-// http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
 package migrate
 
 import (
-	"os"
+	"encoding/json"
 	"path/filepath"
 	"testing"
 
 	testUtils "github.com/perses/perses/internal/test"
 	"github.com/perses/perses/pkg/model/api/config"
+	v1 "github.com/perses/perses/pkg/model/api/v1"
 	"github.com/stretchr/testify/assert"
-	"github.com/stretchr/testify/require"
 )
 
 const testDataFolder = "testdata"
 
-func TestMigrate(t *testing.T) {
+func unmarshalStrictPersesDashboard(data []byte) *v1.Dashboard {
+	result := &v1.Dashboard{}
+	if err := json.Unmarshal(data, result); err != nil {
+		panic(err)
+	}
+	return result
+}
 
+func TestMig_Migrate(t *testing.T) {
 	testSuite := []struct {
 		title                       string
 		inputGrafanaDashboardFile   string
@@ -88,42 +82,37 @@ func TestMigrate(t *testing.T) {
 			expectedPersesDashboardFile: "piechart_perses_dashboard.json",
 			expectedErrorStr:            "",
 		},
+		{
+			title:                       "dashboard from user",
+			inputGrafanaDashboardFile:   "grafana_user_grafana_dashboard.json",
+			expectedPersesDashboardFile: "grafana_user_perses_dashboard.json",
+			expectedErrorStr:            "",
+		},
 	}
+	projectPath := testUtils.GetRepositoryPath()
 
 	for _, test := range testSuite {
 		t.Run(test.title, func(t *testing.T) {
 			inputGrafanaDashboardRaw := testUtils.ReadFile(filepath.Join(testDataFolder, test.inputGrafanaDashboardFile))
 			expectedPersesDashboardRaw := testUtils.ReadFile(filepath.Join(testDataFolder, test.expectedPersesDashboardFile))
-
-			projectPath := testUtils.GetRepositoryPath()
-			svc, err := New(config.Schemas{
-				// use the real schemas for these tests
+			m, err := New(config.Schemas{
 				PanelsPath:    filepath.Join(projectPath, "cue", config.DefaultPanelsPath),
 				QueriesPath:   filepath.Join(projectPath, "cue", config.DefaultQueriesPath),
 				VariablesPath: filepath.Join(projectPath, "cue", config.DefaultVariablesPath),
 			})
-			assert.NoError(t, err)
-
-			actualPersesDashboard, err := svc.Migrate(inputGrafanaDashboardRaw)
-
-			actualErrorStr := ""
 			if err != nil {
-				actualErrorStr = err.Error()
+				t.Fatal(err)
 			}
-			assert.Equal(t, test.expectedErrorStr, actualErrorStr)
-
-			actualPersesDashboardRaw := testUtils.JSONMarshalStrict(actualPersesDashboard)
-			require.JSONEq(t, string(expectedPersesDashboardRaw), string(actualPersesDashboardRaw))
+			grafanaDashboard := &SimplifiedDashboard{}
+			if unmarshallErr := json.Unmarshal(inputGrafanaDashboardRaw, grafanaDashboard); unmarshallErr != nil {
+				t.Fatal(unmarshallErr)
+			}
+			dashboard, err := m.Migrate(grafanaDashboard)
+			if err != nil {
+				t.Fatal(err)
+			}
+			raw := testUtils.JSONMarshalStrict(dashboard)
+			assert.JSONEq(t, string(expectedPersesDashboardRaw), string(raw))
 		})
 	}
-}
-
-func TestRearrangeGrafanaPanelsWithinExpandedRows(t *testing.T) {
-	input, _ := os.ReadFile("testdata/expanded_rows_before.json")
-	expectedAfter, _ := os.ReadFile("testdata/expanded_rows_after.json")
-
-	t.Run("It should move any wrongly-orphaned panels into the right expanded row", func(t *testing.T) {
-		actualAfter := rearrangeGrafanaPanelsWithinExpandedRows(input)
-		require.JSONEq(t, string(expectedAfter), string(actualAfter))
-	})
 }
