+++ internal/api/migrate/testdata/old_grafana_query_perses_dashboard.json
@@ -15,6 +15,9 @@
       {
         "kind": "ListVariable",
         "spec": {
+          "display": {
+            "hidden": false
+          },
           "defaultValue": "argos-world",
           "allowAllValue": false,
           "allowMultiple": false,
