--- internal/api/migratev2/testdata/barchart_grafana_dashboard.json
@@ -1,86 +1,0 @@
-{
-  "editable": false,
-  "graphTooltip": 0,
-  "id": null,
-  "panels": [
-      {
-          "datasource": "$datasource",
-          "description": "Just displaying some metrics with Bar panel",
-          "fieldConfig": {
-              "defaults": {
-                  "mappings": [],
-                  "thresholds": {
-                      "mode": "absolute",
-                      "steps": [
-                          {
-                              "color": "green",
-                              "value": null
-                          },
-                          {
-                              "color": "red",
-                              "value": 80
-                          }
-                      ]
-                  },
-                  "color": {
-                      "mode": "thresholds"
-                  },
-                  "max": 100,
-                  "min": 0,
-                  "unit": "percent",
-                  "decimal": 2
-              },
-              "overrides": []
-          },
-          "gridPos": {
-              "h": 4,
-              "w": 6,
-              "x": 6,
-              "y": 11
-          },
-          "id": 11,
-          "options": {
-              "reduceOptions": {
-                  "values": false,
-                  "calcs": [
-                      "lastNotNull"
-                  ],
-                  "fields": ""
-              },
-              "orientation": "horizontal",
-              "displayMode": "gradient",
-              "valueMode": "color",
-              "namePlacement": "auto",
-              "showUnfilled": true,
-              "sizing": "auto",
-              "minVizWidth": 0,
-              "minVizHeight": 10,
-              "maxVizHeight": 300
-          },
-          "pluginVersion": "11.0.0",
-          "targets": [
-              {
-                  "expr": "up{env=\"demo\",instance=~\"demo.do.prometheus.io.*\"}",
-                  "refID": "A"
-              }
-          ],
-          "title": "Some metrics",
-          "type": "bargauge"
-      }
-  ],
-  "refresh": "1m",
-  "schemaVersion": 25,
-  "style": "dark",
-  "tags": [],
-  "templating": {
-      "list": []
-  },
-  "time": {
-      "from": "now-1h",
-      "to": "now"
-  },
-  "timepicker": {},
-  "timezone": "",
-  "title": "Barchart panel dash",
-  "uid": "barchart"
-}
