--- internal/api/migratev2/testdata/empty_panels_list_perses_dashboard.json
@@ -1,18 +1,0 @@
-{
-  "kind":"Dashboard", 
-  "metadata": {
-    "createdAt":"0001-01-01T00:00:00Z", 
-    "name":"ddgn9k36v60owc", 
-    "project":"", 
-    "updatedAt":"0001-01-01T00:00:00Z", 
-    "version":0
-  }, 
-  "spec": {
-    "display":{
-      "name":"test dashboard"
-    }, 
-    "duration":"1h", 
-    "layouts":[], 
-    "panels":{}
-  }
-}
