--- internal/api/migratev2/testdata/old_grafana_query_perses_dashboard.json
@@ -1,107 +1,0 @@
-{
-  "kind": "Dashboard",
-  "metadata": {
-    "name": "perses-testing-old-query",
-    "createdAt": "0001-01-01T00:00:00Z",
-    "updatedAt": "0001-01-01T00:00:00Z",
-    "version": 0,
-    "project": ""
-  },
-  "spec": {
-    "display": {
-      "name": "Perses testing / old query format"
-    },
-    "variables": [
-      {
-        "kind": "ListVariable",
-        "spec": {
-          "display": {
-            "hidden": false
-          },
-          "defaultValue": "argos-world",
-          "allowAllValue": false,
-          "allowMultiple": false,
-          "plugin": {
-            "kind": "StaticListVariable",
-            "spec": {
-              "values": [
-                "grafana",
-                "migration",
-                "not",
-                "supported"
-              ]
-            }
-          },
-          "name": "datasource"
-        }
-      },
-      {
-        "kind": "ListVariable",
-        "spec": {
-          "display": {
-            "name": "Paas",
-            "hidden": false
-          },
-          "defaultValue": "prd1",
-          "allowAllValue": false,
-          "allowMultiple": true,
-          "sort": "alphabetical-asc",
-          "plugin": {
-            "kind": "PrometheusPromQLVariable",
-            "spec": {
-              "expr": "max by (stack) (thanos_build_info{prometheus=\"argos\",federation=\"\"}) or ( label_replace(vector(1), \"stack\", \"external\", \"\", \"\") ) ",
-              "labelName": "stack"
-            }
-          },
-          "name": "stack"
-        }
-      }
-    ],
-    "panels": {
-      "0": {
-        "kind": "Panel",
-        "spec": {
-          "display": {
-            "name": "Panel Title"
-          },
-          "plugin": {
-            "kind": "TimeSeriesChart",
-            "spec": {
-              "legend": {
-                "mode": "list",
-                "position": "bottom",
-                "values": []
-              },
-              "visual": {
-                "areaOpacity": 0,
-                "connectNulls": false,
-                "display": "line",
-                "lineWidth": 3,
-                "stack": "all"
-              }
-            }
-          }
-        }
-      }
-    },
-    "layouts": [
-      {
-        "kind": "Grid",
-        "spec": {
-          "items": [
-            {
-              "x": 0,
-              "y": 0,
-              "width": 12,
-              "height": 8,
-              "content": {
-                "$ref": "#/spec/panels/0"
-              }
-            }
-          ]
-        }
-      }
-    ],
-    "duration": "1h"
-  }
-}
