--- internal/api/migratev2/testdata/piechart_grafana_dashboard.json
@@ -1,86 +1,0 @@
-{
-    "annotations": {},
-    "editable": true,
-    "fiscalYearStartMonth": 0,
-    "graphTooltip": 0,
-    "id": 937,
-    "links": [],
-    "panels": [
-        {
-            "fieldConfig": {
-                "defaults": {
-                    "color": {
-                        "mode": "palette-classic"
-                    },
-                    "custom": {
-                        "hideFrom": {
-                            "legend": false,
-                            "tooltip": false,
-                            "viz": false
-                        }
-                    },
-                    "mappings": []
-                },
-                "overrides": []
-            },
-            "gridPos": {
-                "h": 6,
-                "w": 12,
-                "x": 12,
-                "y": 11
-            },
-            "id": 1,
-            "options": {
-                "legend": {
-                    "displayMode": "list",
-                    "placement": "bottom",
-                    "showLegend": true
-                },
-                "pieType": "pie",
-                "reduceOptions": {
-                    "calcs": [
-                        "lastNotNull"
-                    ],
-                    "fields": "",
-                    "values": false
-                },
-                "tooltip": {
-                    "mode": "single",
-                    "sort": "none"
-                }
-            },
-            "pluginVersion": "11.3.0-pre",
-            "targets": [
-                {
-                    "expr": "matching{job=\"abc\"}",
-                    "legendFormat": "matching roles",
-                    "refId": "A"
-                },
-                {
-                    "expr": "nonmatching{job=\"abc2\"}",
-                    "legendFormat": "non-matching roles",
-                    "refId": "B"
-                }
-            ],
-            "title": "Piechart sample",
-            "type": "piechart"
-        }
-    ],
-    "preload": false,
-    "refresh": "1m",
-    "schemaVersion": 40,
-    "tags": [],
-    "templating": {
-        "list": []
-    },
-    "time": {
-        "from": "now-1h",
-        "to": "now"
-    },
-    "timepicker": {},
-    "timezone": "",
-    "title": "Piechart sample",
-    "uid": "piechart",
-    "version": 1,
-    "weekStart": ""
-}
