--- internal/api/migratev2/testdata/stat_calc_undefined_perses_dashboard.json
@@ -1,72 +1,0 @@
-{
-  "kind": "Dashboard",
-  "metadata": {
-    "name": "stat-undefined-calc",
-    "createdAt": "0001-01-01T00:00:00Z",
-    "updatedAt": "0001-01-01T00:00:00Z",
-    "version": 0,
-    "project": ""
-  },
-  "spec": {
-    "display": {
-      "name": "Stat panel with undefined reduceOptions"
-    },
-    "panels": {
-      "0": {
-        "kind": "Panel",
-        "spec": {
-          "display": {
-            "name": "Some metrics",
-            "description": "Just displaying some metrics with Stat panel"
-          },
-          "plugin": {
-            "kind": "StatChart",
-            "spec": {
-              "calculation": "last",
-              "thresholds": {
-                "steps": [
-                  {
-                    "color": "blue",
-                    "value": 0
-                  }
-                ]
-              }
-            }
-          },
-          "queries": [
-            {
-              "kind": "TimeSeriesQuery",
-              "spec": {
-                "plugin": {
-                  "kind": "PrometheusTimeSeriesQuery",
-                  "spec": {
-                    "query": "up{env=\"demo\",instance=~\"demo.do.prometheus.io.*\"}"
-                  }
-                }
-              }
-            }
-          ]
-        }
-      }
-    },
-    "layouts": [
-      {
-        "kind": "Grid",
-        "spec": {
-          "items": [
-            {
-              "x": 6,
-              "y": 11,
-              "width": 6,
-              "height": 4,
-              "content": {
-                "$ref": "#/spec/panels/0"
-              }
-            }
-          ]
-        }
-      }
-    ],
-    "duration": "1h"
-  }
-}
